refactor: drawer -> data-bubbles
