+++ src/entities/data-bubbles/lib/DrawableDataBubble.ts
