+++ src/entities/data-bubbles/lib/drawBuble.ts
