+++ src/entities/data-bubbles/lib/getFunctionGetColorByDelta.ts
