+++ src/entities/data-bubbles/ui/CanvasDataBubbles/index.tsx
