+++ src/widgets/WidgetSectionShare/ui/BoxDrawerImageDownload/index.tsx
@@ -1,4 +1,4 @@
-import type { DrawerDataBubbles } from "@/entities/drawer";
+import type { DrawerDataBubbles } from "@/entities/data-bubbles";
 import { download } from "@/shared/lib/files/download";
 import { getFileExtensionByMimeType } from "@/shared/lib/files/getFileExtensionByMimeType";
 import { useEffect, useState } from "react";
